Move openPostInReaderOrInAppWebView from StatsUtils to new helper object
